Fix url typo in DruidNavigator and refactor
Wiki enums are now only used in testing
