+++ fili-generic-example/src/test/groovy/com/yahoo/wiki/webservice/web/endpoints/AutomaticDruidConfigLoaderSpec.groovy
@@ -50,6 +50,7 @@ public class AutomaticDruidConfigLoaderSpec extends Specification {
                     "/?full") {
                 return expectedMetricsAndDimensions
             }
+            return "BAD ERROR WHAT HAPPENED"
         }
     }
